modified to manage the activities at the window level
Timan Rebel has discovered the best technique I've seen for accurate handling of paused/resumed events in Titanium on android.  It requires that the developer use discipline when creating and opening windows.  The technique demands that the app track the start/stop events on every single activity in the app.  So every time you create a window, you have to let this centralized system know about it so that it can monitor the start/stop events.  You also have to give every activity a unique name.

My changes are trying to make that work a little bit easier.  The developer doesn't have to name the activities; the Context library will handle that for you internally.  And the developer doesn't even have to think about the activity associated with each window; he just calls Context.track(), passing the window object.